pyproject.toml
[project]
name = "ollama-deepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "firecrawl-py>=1.16.0",
    "ollama>=0.4.8",
    "pydantic>=2.11.3",
    "tqdm>=4.67.1",
]
